+++ src/Auth/Auth.php
@@ -72,11 +72,17 @@ class Auth implements AuthInterface
      */
     protected function getUsersByField(string $field, iterable $values): array
     {
+        $values = (array) $values;
+
+        if (count($values) === 0) {
+            return [];
+        }
+
         $response = $this->internalApi->call(
             Component::CORE,
             '/auth/users/from/' . $field,
             [
-                $field => (array)$values
+                $field => $values
             ]
         );